intelligent_placement_prep_analyser/backend/requirements.txt

fastapi
uvicorn
python-jose[cryptography]
argon2-cffi
sqlalchemy
psycopg2-binary
langchain
langchain-community
langchain-google-genai
google-generativeai
chromadb
faiss-cpu
python-multipart
speechrecognition
pydub
python-dotenv
pdfplumber
pydantic[email]
pydantic-settings
sentence-transformers
cryptography
